requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
SQLAlchemy==2.0.32
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==0.21.1
pydantic==2.8.2
pydantic_core==2.20.1